Fixed runtime errors.
