+++ utils/gtpchar2num.js
@@ -1,0 +1,7 @@
+function gtpchar2num(ch) {
+    if (ch === "." || !ch)
+        return -1;
+    return "abcdefghjklmnopqrstuvwxyz".indexOf(ch.toLowerCase());
+}
+
+exports.gtpchar2num = gtpchar2num;
