+++ utils/num2char.js
@@ -1,0 +1,6 @@
+function num2char(num) {
+    if (num === -1) return ".";
+    return "abcdefghijklmnopqrstuvwxyz"[num];
+}
+
+exports.num2char = num2char;
